requirements.txt
msedge-selenium-tools==3.141.4
selenium==3.141.0
urllib3==1.26.8
python-dotenv==0.19.2